Fixed the returning from functions issue.
The program did not return properly because the function that executes blocks are used for both if/for and functions.
We need to stop returning after we're out of a function call. I made a new run_block function for the if/for functions.
There are still check for incorrect scoping and bad expresssion on gradescope that I need to resolve. My file has passed all
local cases, including the ones imported from campuswire.